Refactor popup models into separate files
Moved EULA and game server list models from Popup_EULA.xaml.cs and Popup_Services.xaml.cs into dedicated files under Models/EULA and Models/Services. Updated references to use the new model classes, improving code organization and maintainability.
