+++ launcher/Views/Popups/Models/EULA/Data.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace launcher.Views.Popups.Models.EULA
+{
+    public class Data
+    {
+        public int version { get; set; }
+        public string lang { get; set; }
+        public string contents { get; set; }
+        public DateTime modified { get; set; }
+        public string language { get; set; }
+    }
+}
