+++ launcher/Views/Popups/Models/EULA/EULA.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace launcher.Views.Popups.Models.EULA
+{
+    public class EULA
+    {
+        public bool success { get; set; }
+        public Data data { get; set; }
+    }
+}
