+++ launcher/Views/Popups/Models/Services/GameServerList.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace launcher.Views.Popups.Models.Services
+{
+    public class GameServerList
+    {
+        public bool success { get; set; }
+        public List<Server> servers { get; set; }
+    }
+}
