+++ launcher/Views/Popups/Popup_Services.xaml.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using launcher.Views.Popups.Models.Services;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -202,27 +203,5 @@ namespace launcher
         {
             Process.Start(new ProcessStartInfo("cmd", $"/c start {status_url}") { CreateNoWindow = true });
         }
-    }
-
-    public class GameServerList
-    {
-        public bool success { get; set; }
-        public List<Server> servers { get; set; }
-    }
-
-    public class Server
-    {
-        public string maxPlayers { get; set; }
-        public string port { get; set; }
-        public string checksum { get; set; }
-        public string name { get; set; }
-        public string ip { get; set; }
-        public string description { get; set; }
-        public string hidden { get; set; }
-        public string playerCount { get; set; }
-        public string playlist { get; set; }
-        public string key { get; set; }
-        public string region { get; set; }
-        public string map { get; set; }
     }
 }
